% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseConverters.R
\name{bin2dec}
\alias{bin2dec}
\title{Binary to Decimal converter
Will not work on binary numbers longer than 54bits. R integer maxes out at .Machine$integer.max, but
R will sum up to 18014398509481984, or 54bits, as long as it is not stored as an integer object.}
\usage{
bin2dec(bin)
}
\arguments{
\item{bin}{A single binary number as a string with nchar(bin) <= 54}
}
\value{
A single decimal number as a string
}
\description{
Binary to Decimal converter
Will not work on binary numbers longer than 54bits. R integer maxes out at .Machine$integer.max, but
R will sum up to 18014398509481984, or 54bits, as long as it is not stored as an integer object.
}
\examples{
bin2dec('1001100')
}
