% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseConverters.R
\name{bin2hex}
\alias{bin2hex}
\title{Binary to Hexadecimal converter}
\usage{
bin2hex(bin)
}
\arguments{
\item{bin}{A single binary number as a string}
}
\value{
A single hexadecimal number as a string
}
\description{
Binary to Hexadecimal converter
}
\examples{
bin2hex('1001101')
}
