% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_converters.R
\name{binary_to_decimal}
\alias{binary_to_decimal}
\title{Binary to Decimal converter}
\usage{
binary_to_decimal(bin)
}
\arguments{
\item{bin}{A single binary number as a string - nchar(bin) <= 54}
}
\value{
A single decimal number as a string
}
\description{
Binary to Decimal converter
}
\examples{
binary_to_decimal('1001100')
}
