% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_converters.R
\name{binary_to_hexadecimal}
\alias{binary_to_hexadecimal}
\title{Binary to Hexadecimal converter}
\usage{
binary_to_hexadecimal(bin)
}
\arguments{
\item{bin}{A single binary number as a string}
}
\value{
A single hexadecimal number as a string
}
\description{
Binary to Hexadecimal converter
}
\examples{
binary_to_hexadecimal('1001101')
}
