% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_all}
\alias{convert_to_all}
\title{Convert a list of codes to all other formats}
\usage{
convert_to_all(.data)
}
\arguments{
\item{.data}{A vector of character strings}
}
\value{
A tibble containing the original codes with a detected format column plus a column for each known format
}
\description{
Convert a list of codes to all other formats
}
\examples{
convert_to_all(c('3DD.ABC4567890'))
}
