% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_iso64bitr}
\alias{convert_to_iso64bitr}
\title{Convert any recognised ISO11784/5 format to Raw Hexadecimal format, animal tag bit on the Right}
\usage{
convert_to_iso64bitr(.data, format = NULL)
}
\arguments{
\item{.data}{A vector of character strings}

\item{format}{a string indicating which format the ID is in, intended to be used with output from get_iso11784_format
c('unknown', 'isodecimal', 'isodothex','iso64bitl', 'iso64bitr')}
}
\value{
A converted vector of character strings
}
\description{
Convert any recognised ISO11784/5 format to Raw Hexadecimal format, animal tag bit on the Right
}
\examples{
convert_to_iso64bitr(c('999012345678901','8000F9C2DFDC1C36','3E7.02DFDC1C37'))
}
