% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseConverters.R
\name{dec2bin}
\alias{dec2bin}
\title{Decimal to Binary converter}
\usage{
dec2bin(dec)
}
\arguments{
\item{dec}{A single decimal number as a string}
}
\value{
A single binary number as a string
}
\description{
Decimal to Binary converter
}
\examples{
dec2bin('1234')
}
