% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseConverters.R
\name{dec2hex}
\alias{dec2hex}
\title{Decimal to Hexadecimal converter}
\usage{
dec2hex(dec)
}
\arguments{
\item{dec}{A single decimal number as a string}
}
\value{
A single hexadecimal number as a string
}
\description{
Decimal to Hexadecimal converter
}
\examples{
dec2hex('1234')
}
