% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_converters.R
\name{decimal_to_binary}
\alias{decimal_to_binary}
\title{Decimal to Binary converter}
\usage{
decimal_to_binary(dec)
}
\arguments{
\item{dec}{A single decimal number as a string}
}
\value{
A single binary number as a string
}
\description{
Decimal to Binary converter
}
\examples{
decimal_to_binary('1234')
}
