% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_converters.R
\name{decimal_to_hexadecimal}
\alias{decimal_to_hexadecimal}
\title{Decimal to Hexadecimal converter}
\usage{
decimal_to_hexadecimal(dec)
}
\arguments{
\item{dec}{A single decimal number as a string}
}
\value{
A single hexadecimal number as a string
}
\description{
Decimal to Hexadecimal converter
}
\examples{
decimal_to_hexadecimal('1234')
}
