% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseConverters.R
\name{hex2bin}
\alias{hex2bin}
\title{Hexadecimal to Binary converter
This is simply a wrapper for dec2bin and hex2dec}
\usage{
hex2bin(hex)
}
\arguments{
\item{hex}{A single hexadecimal number as a string}
}
\value{
A single binary number as a string
}
\description{
Hexadecimal to Binary converter
This is simply a wrapper for dec2bin and hex2dec
}
\examples{
hex2bin('ABC123')
}
