% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseConverters.R
\name{hex2dec}
\alias{hex2dec}
\title{Base Converters
These functions are used by the main ISO11784 converters to switch between bases 2, 10 and 16
Hexadecimal to Decimal converter}
\usage{
hex2dec(hex)
}
\arguments{
\item{hex}{A single hexadecimal number as a string. Max '40000000000000'}
}
\value{
A single decimal number as a string
}
\description{
Base Converters
These functions are used by the main ISO11784 converters to switch between bases 2, 10 and 16
Hexadecimal to Decimal converter
}
\examples{
hex2dec('ABC123')
}
