% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_converters.R
\name{hexadecimal_to_binary}
\alias{hexadecimal_to_binary}
\title{Hexadecimal to Binary converter
This is simply a wrapper for decimal_to_binary and hexadecimal_to_decimal}
\usage{
hexadecimal_to_binary(hex)
}
\arguments{
\item{hex}{A single hexadecimal number as a string}
}
\value{
A single binary number as a string
}
\description{
Hexadecimal to Binary converter
This is simply a wrapper for decimal_to_binary and hexadecimal_to_decimal
}
\examples{
hexadecimal_to_binary('ABC123')
}
