% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_converters.R
\name{hexadecimal_to_decimal}
\alias{hexadecimal_to_decimal}
\title{Hexadecimal to Decimal converter}
\usage{
hexadecimal_to_decimal(hex)
}
\arguments{
\item{hex}{A single hexadecimal number as a string. Max '40000000000000'}
}
\value{
A single decimal number as a string
}
\description{
Hexadecimal to Decimal converter
}
\examples{
hexadecimal_to_decimal('ABC123')
}
