% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_isodecimal_tform.R
\name{iso64bitleft_to_isodecimal}
\alias{iso64bitleft_to_isodecimal}
\title{ISO 11784 Raw Hexadecimal format, animal ID on the LEFT, To ISO11784 15 Digit Decimal}
\usage{
iso64bitleft_to_isodecimal(.data)
}
\arguments{
\item{.data}{A vector of Raw Hexadecimal format strings.}
}
\value{
A vector of ISO11784 15 Digit Decimal format strings.
}
\description{
ISO 11784 Raw Hexadecimal format, animal ID on the LEFT, To ISO11784 15 Digit Decimal
}
\examples{
iso64bitleft_to_isodecimal(c('8000ABCDEF123456'))
}
