% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_isodecimal_tform.R
\name{ISO64bitLeftToISODecimal}
\alias{ISO64bitLeftToISODecimal}
\title{ISO 11784 Raw Hexadecimal format, animal ID on the LEFT, To ISO11784 15 Digit Decimal}
\usage{
ISO64bitLeftToISODecimal(.data)
}
\arguments{
\item{.data}{A vector of Raw Hexadecimal format strings.}
}
\value{
A vector of ISO11784 15 Digit Decimal format strings.
}
\description{
ISO 11784 Raw Hexadecimal format, animal ID on the LEFT, To ISO11784 15 Digit Decimal
}
\examples{
ISO64bitLeftToISODecimal(c('8000ABCDEF123456'))
}
