% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_isodecimal_tform.R
\name{isodecimal_to_iso64bitright}
\alias{isodecimal_to_iso64bitright}
\title{ISO11784 15 Digit Decimal format To ISO11784 Raw Hexadecimal format, animal ID on the RIGHT}
\usage{
isodecimal_to_iso64bitright(.data)
}
\arguments{
\item{.data}{A vector of ISO11784 15 Digit Decimal format strings.}
}
\value{
A vector of Raw Hexadecimal format strings.
}
\description{
ISO11784 15 Digit Decimal format To ISO11784 Raw Hexadecimal format, animal ID on the RIGHT
}
\examples{
isodecimal_to_iso64bitright(c('989737733408912'))
}
