% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_isodecimal_tform.R
\name{isodecimal_to_isodothex}
\alias{isodecimal_to_isodothex}
\title{ISO11784 15 Digit Decimal format To ISO11784 Dot Hexadecimal}
\usage{
isodecimal_to_isodothex(.data)
}
\arguments{
\item{.data}{A vector of ISO11784 15 Digit Decimal format strings.}
}
\value{
A vector of Dot Hexadecimal format strings.
}
\description{
ISO11784 15 Digit Decimal format To ISO11784 Dot Hexadecimal
}
\examples{
isodecimal_to_isodothex(c('989737733408912'))
}
