% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_isodecimal_tform.R
\name{ISODecimalToISO64bitLeft}
\alias{ISODecimalToISO64bitLeft}
\title{ISO11784 15 Digit Decimal format To ISO11784 Raw Hexadecimal format, animal ID on the LEFT}
\usage{
ISODecimalToISO64bitLeft(.data)
}
\arguments{
\item{.data}{A vector of ISO11784 15 Digit Decimal format strings.}
}
\value{
A vector of Raw Hexadecimal format strings.
}
\description{
ISO11784 15 Digit Decimal format To ISO11784 Raw Hexadecimal format, animal ID on the LEFT
}
\examples{
ISODecimalToISO64bitLeft(c('989737733408912'))
}
