% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromISODecimalConverters.R
\name{ISODecimalToISO64bitRight}
\alias{ISODecimalToISO64bitRight}
\title{ISO11784 15 Digit Decimal format To ISO11784 Raw Hexadecimal format, animal ID on the RIGHT}
\usage{
ISODecimalToISO64bitRight(.data)
}
\arguments{
\item{.data}{A vector of ISO11784 15 Digit Decimal format strings.}
}
\value{
A vector of Raw Hexadecimal format strings.
}
\description{
ISO11784 15 Digit Decimal format To ISO11784 Raw Hexadecimal format, animal ID on the RIGHT
}
\examples{
ISODecimalToISO64bitRight(c('989737733408912'))
}
