% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_isodecimal_tform.R
\name{ISOdecimalToISOdothex}
\alias{ISOdecimalToISOdothex}
\title{ISO11784 15 Digit Decimal format To ISO11784 Dot Hexadecimal}
\usage{
ISOdecimalToISOdothex(.data)
}
\arguments{
\item{.data}{A vector of ISO11784 15 Digit Decimal format strings.}
}
\value{
A vector of Dot Hexadecimal format strings.
}
\description{
ISO11784 15 Digit Decimal format To ISO11784 Dot Hexadecimal
}
\examples{
ISOdecimalToISOdothex(c('989737733408912'))
}
