% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_isodecimal_tform.R
\name{isodothex_to_isodecimal}
\alias{isodothex_to_isodecimal}
\title{ISO 11784 Dot Hexadecimal format To ISO11784 15 Digit Decimal}
\usage{
isodothex_to_isodecimal(.data)
}
\arguments{
\item{.data}{A vector of Dot Hexadecimal format strings.}
}
\value{
A vector of ISO11784 15 Digit Decimal format strings.
}
\description{
ISO 11784 Dot Hexadecimal format To ISO11784 15 Digit Decimal
}
\examples{
isodothex_to_isodecimal(c('3DD.ABC4567890'))
}
