% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_user_functions.R
\name{convert_to_iso64bitr}
\alias{convert_to_iso64bitr}
\title{Convert any recognised ISO11784/5 format to Raw Hexadecimal format, animal tag bit on the Right}
\usage{
convert_to_iso64bitr(.data)
}
\arguments{
\item{.data}{A vector of character strings}
}
\value{
A converted vector of character strings
}
\description{
Convert any recognised ISO11784/5 format to Raw Hexadecimal format, animal tag bit on the Right
}
\examples{
convert_to_iso64bitr(c('999012345678901','8000F9C2DFDC1C36','3E7.02DFDC1C37'))
}
