% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_user_functions.R
\name{convert_to_isodothex}
\alias{convert_to_isodothex}
\title{Convert any recognised ISO11784/5 format to Dot Hexadecimal format}
\usage{
convert_to_isodothex(.data)
}
\arguments{
\item{.data}{A vector of character strings}
}
\value{
A converted vector of character strings
}
\description{
Convert any recognised ISO11784/5 format to Dot Hexadecimal format
}
\examples{
convert_to_isodothex(c('999012345678901','8000F9C2DFDC1C36','EC383BFB439F0001'))
}
