% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromISODecimalConverters.R
\name{ISOdecimalToISOdothex}
\alias{ISOdecimalToISOdothex}
\title{Main functions for converting from ISO11784 15 Digit Decimal to other formats
These are wrappers for the functions in baseConverters.R
ISO11784 15 Digit Decimal format To ISO11784 Dot Hexadecimal}
\usage{
ISOdecimalToISOdothex(dec)
}
\arguments{
\item{dec}{A vector of ISO11784 15 Digit Decimal format strings.}
}
\value{
A vector of Dot Hexadecimal format strings.
}
\description{
Main functions for converting from ISO11784 15 Digit Decimal to other formats
These are wrappers for the functions in baseConverters.R
ISO11784 15 Digit Decimal format To ISO11784 Dot Hexadecimal
}
\examples{
ISOdecimalToISOdothex(c('989737733408912'))
}
